refactor(api): move mission enrichment admin logic
